Revert "Update Effekseer and refactor"
This reverts commit 96f24dc99c11f82e920f0b2676a036bdbcf206c0.

+++ Plugins/Effekseer/Source/Effekseer/Private/Effekseer/Effekseer/Material/Effekseer.CompiledMaterial.cpp
@@ -21,27 +21,27 @@ public:
 
 	void SetVertexShaderData(MaterialShaderType type, const std::vector<uint8_t>& data)
 	{
-		vertexShaders_.at(static_cast<size_t>(type)) = data;
+		vertexShaders_.at(static_cast<int>(type)) = data;
 	}
 
 	void SetPixelShaderData(MaterialShaderType type, const std::vector<uint8_t>& data)
 	{
-		pixelShaders_.at(static_cast<size_t>(type)) = data;
+		pixelShaders_.at(static_cast<int>(type)) = data;
 	}
 
 	const uint8_t* GetVertexShaderData(MaterialShaderType type) const override
 	{
-		return vertexShaders_.at(static_cast<size_t>(type)).data();
+		return vertexShaders_.at(static_cast<int>(type)).data();
 	}
 
 	int32_t GetVertexShaderSize(MaterialShaderType type) const override
 	{
-		return static_cast<int32_t>(vertexShaders_.at(static_cast<size_t>(type)).size());
+		return static_cast<int32_t>(vertexShaders_.at(static_cast<int>(type)).size());
 	}
 
 	const uint8_t* GetPixelShaderData(MaterialShaderType type) const override
 	{
-		return pixelShaders_.at(static_cast<size_t>(type)).data();
+		return pixelShaders_.at(static_cast<int>(type)).data();
 	}
 
 	int32_t GetPixelShaderSize(MaterialShaderType type) const override
